move block sorting to block.rs
